moved update protocol into the pages folder
